requirements.txt
requests>=2.25.0
websocket-client>=1.0.0
websockets>=10.0
tiktoken>=0.5.0
pyyaml>=6.0
